Separate CAN adapters from utils
